test(spur-net): integrate WireGuard mesh tests into native_host/e2e
Fold the WireGuard mesh E2E suite into the existing native_host/e2e suite
instead of a parallel directory. The WG fixtures now build on the base
unstarted_cluster + config-override path (same as the k8s fixtures): they
enable network.wg_enabled in config and stand up the spur0 mesh between
provision and start. Session scaffolding that duplicated the base conftest
is dropped.

WG tests run automatically wherever the nodes allow — no SPUR_TEST_WG opt-in.
The fixtures install WireGuard on a node that lacks it (wireguard-tools, then
the kernel module or wireguard-go userspace fallback) and skip only where a
data plane genuinely can't be provided, mirroring the capability-gated
gpu_cluster / k8s_multicp_cluster fixtures. Setup and teardown are per test:
the mesh + k0s are stood up and fully torn down (spur0 removed, conf deleted,
k0s reset) so a shared node is left clean for the next, possibly non-WG, test.

Tests are split by topic to match the base convention: test_wg_mesh.py (raw
mesh bring-up) and test_wg_k0s.py (k0s-over-mesh). CI drops the separate path
and the WG-install step (the fixture handles it); docs updated.
