+++ tests/native_host/e2e/conftest.py
@@ -7,13 +7,16 @@ Pytest configuration and fixtures for Spur native-host E2E tests.
 See docs/developer/building.rst for full environment variable reference.
 """
 
+import logging
 import os
-import time
 from pathlib import Path
 
 import pytest
 
 from cluster import SshNode, SpurCluster, deep_merge, ensure_bins, make_remote_dir
+from wg_cluster import WG_IFACE, WgMesh, wg_available
+
+logger = logging.getLogger(__name__)
 
 _REPO_ROOT = Path(__file__).resolve().parents[3]
 
@@ -35,6 +38,11 @@ def pytest_configure(config):
     config.addinivalue_line(
         "markers",
         "gpu: requires GPU hardware on the nodes (auto-applied to tests using GPU fixtures)",
+    )
+    config.addinivalue_line(
+        "markers",
+        "wireguard: WireGuard mesh tests (fixture installs wg + stands up a real "
+        "spur0 mesh; rootful; auto-skips only where a data plane can't be provided)",
     )
 
 
@@ -397,3 +405,232 @@ def mpi_cluster(ssh_nodes, remote_bin_dir, cluster_config_overrides):
     c.mpi_preflight(1)
     yield c
     c.teardown()
+
+
+# --- WireGuard mesh fixtures -------------------------------------------------
+#
+# WireGuard tests run automatically wherever the nodes allow a real mesh — there
+# is no opt-in env var (mirroring gpu_cluster / k8s_multicp_cluster, which also
+# auto-run on capable beds). The fixtures ENABLE WireGuard rather than detect it:
+# they flip network.wg_enabled on in config AND install the wg data plane on the
+# node if it is missing (_ensure_wireguard), so the tests execute instead of
+# skipping. They skip only on a genuine environmental impossibility (no kernel
+# module AND no wireguard-go userspace fallback obtainable) or missing sudo.
+
+
+_WG_CIDR = "10.44.0.0/16"
+
+
+def _wg_addresses() -> dict[int, str] | None:
+    """Optional explicit per-node mesh addresses, positional like SPUR_TEST_NODES.
+
+    Mirrors spur-toolkit's per-node ``spur_wg_address`` inventory var. Example:
+    ``SPUR_TEST_WG_ADDRESSES=10.44.0.1,10.44.0.2,10.44.0.3``. When unset, WgMesh
+    falls back to its default assignment. When set, the count must match
+    SPUR_TEST_NODES.
+    """
+    raw = os.environ.get("SPUR_TEST_WG_ADDRESSES", "").strip()
+    if not raw:
+        return None
+    addrs = [a.strip() for a in raw.split(",") if a.strip()]
+    nodes = _get_nodes_config()
+    if len(addrs) != len(nodes):
+        pytest.exit(
+            f"SPUR_TEST_WG_ADDRESSES has {len(addrs)} entries but SPUR_TEST_NODES "
+            f"has {len(nodes)} — they must match", returncode=1
+        )
+    return {i: a for i, a in enumerate(addrs)}
+
+
+def _ensure_wireguard(cluster: SpurCluster) -> None:
+    """Make every node able to run a real WireGuard mesh, installing if needed.
+
+    Enables WireGuard even on a node that starts without it: installs
+    ``wireguard-tools`` when ``wg``/``wg-quick`` are missing, then ensures a data
+    plane — the in-tree kernel module (``modprobe wireguard``) or, if that won't
+    load, the ``wireguard-go`` userspace fallback. Only when neither a module nor
+    the userspace fallback can be obtained does it ``pytest.skip`` (a genuine
+    environmental limit, not a policy toggle). apt-based: assumes Debian/Ubuntu
+    nodes (the testbed + CI image).
+    """
+    sudo = cluster._sudo_prefix()
+    for node in cluster.nodes:
+        if "OK" not in node.exec_allow_fail("command -v wg >/dev/null && command -v wg-quick >/dev/null && echo OK"):
+            node.exec_allow_fail(
+                f"{sudo}apt-get update -qq && {sudo}DEBIAN_FRONTEND=noninteractive "
+                "apt-get install -y -qq wireguard-tools 2>/dev/null || true"
+            )
+        # Ensure a data plane: prefer the kernel module, else install wireguard-go.
+        node.exec_allow_fail(
+            f"{sudo}modprobe wireguard 2>/dev/null "
+            "|| command -v wireguard-go >/dev/null 2>&1 "
+            f"|| {sudo}DEBIAN_FRONTEND=noninteractive apt-get install -y -qq wireguard-go 2>/dev/null "
+            "|| true"
+        )
+        ok, reason = wg_available(node, sudo)
+        if not ok:
+            pytest.skip(f"WireGuard unavailable on {node.host} and could not be provisioned: {reason}")
+
+
+@pytest.fixture
+def wg_mesh(ssh_nodes, remote_bin_dir):
+    """A raw WireGuard mesh across all nodes (no spur daemons).
+
+    Yields a :class:`WgMesh` with the interface provisioned via ``net init`` on
+    node 0 and ``net join`` on the rest, promoted to a full mesh. The mesh is
+    installed if missing (``_ensure_wireguard``) and fully torn down (interfaces +
+    confs removed) after the test so a shared node is left clean.
+    """
+    if len(ssh_nodes) < 2:
+        pytest.skip(f"WG mesh needs >= 2 nodes (got {len(ssh_nodes)})")
+    c = SpurCluster(ssh_nodes, make_remote_dir(), remote_bin_dir)
+    c.provision()  # resolves node_names; no daemons started
+    c.root_agent_preflight()  # rootful: spur0 needs wg/ip via sudo; skips if none
+    _ensure_wireguard(c)
+    mesh = WgMesh(ssh_nodes, c.node_names, remote_bin_dir, c._sudo_prefix(),
+                  iface=WG_IFACE, wg_addresses=_wg_addresses())
+    indices = list(range(len(ssh_nodes)))
+    try:
+        mesh.bring_up(indices)
+    except Exception:
+        mesh.teardown(indices)
+        c.teardown()
+        raise
+    yield mesh
+    mesh.teardown(indices)
+    c.teardown()
+
+
+def _reset_k0s_all_nodes(c):
+    """Forcibly reset k0s on every node, independent of the controller.
+
+    `spur k8s down --reset` only flags phase=Down; the controller's reconcile
+    loop stops the k0s components *asynchronously*. Because the fixture kills
+    spurctld/spurd right after (c.teardown()), that async reset is cut off — so
+    without this, k0s services and the etcd datadir survive and the NEXT wg_k0s
+    test inherits dirty state (membership won't shrink, the API server refuses
+    connections). Reset each node directly so every test starts from a clean slate.
+    """
+    sudo = c._sudo_prefix()
+    for node in c.nodes:
+        node.exec_allow_fail(
+            f"{sudo}systemctl stop k0scontroller k0sworker 2>/dev/null || true"
+        )
+        node.exec_allow_fail(f"{sudo}k0s stop 2>/dev/null || true")
+        node.exec_allow_fail(f"{sudo}k0s reset 2>/dev/null || true")
+        node.exec_allow_fail(
+            f"{sudo}rm -rf /etc/k0s /var/lib/k0s /run/k0s 2>/dev/null || true"
+        )
+        # Best-effort steps above (|| true) can't wedge teardown — but a reset
+        # that silently failed would resurface as a baffling failure in the NEXT
+        # test that inherits the leftover k0s state. Verify the datadir is gone
+        # and warn (naming the node) if not, so blame lands on the offender.
+        leftover = node.exec_allow_fail(
+            "test -e /var/lib/k0s && echo DIRTY || echo CLEAN"
+        )
+        if "DIRTY" in leftover:
+            logger.warning(
+                "k0s reset left /var/lib/k0s behind on %s — the next wg_k0s test "
+                "may inherit dirty state (stale etcd membership / API refusals)",
+                node.host,
+            )
+
+
+def _deploy_wg_k0s(ssh_nodes, remote_bin_dir, *, control_plane_index: int = 0):
+    """Deploy a rootful, WG-enabled, calico k0s cluster over the mesh.
+
+    Returns a running SpurCluster (daemons up, k0s NOT yet `up`) with the mesh
+    established. ``wg_enabled=true`` only tells spurd to *read* its mesh key/IP
+    from ``spur0`` — it does not create the interface, so the mesh is stood up
+    (``net init``/``join``) before the daemons start, else spurd reports no
+    ``wg_pubkey`` and the reconcile runs meshless.
+
+    The live :class:`WgMesh` is attached as ``c.wg_mesh`` so the fixture can tear
+    it down; ``c.wg_mesh_indices`` records which nodes it covers.
+    """
+    c = SpurCluster(ssh_nodes, make_remote_dir(), remote_bin_dir)
+    c.provision()
+    c.root_agent_preflight()  # skips if no sudo
+    _ensure_wireguard(c)
+
+    # Start from a clean k0s slate: a prior test (or a crashed run) can leave k0s
+    # services + an etcd datadir behind, which corrupts this cluster's membership.
+    _reset_k0s_all_nodes(c)
+
+    # Establish the mesh first (spur0 up on every node) so spurd reads its
+    # wg_pubkey + mesh IP at startup and the reconcile has a real mesh to
+    # converge. bring_up() idempotently pre-cleans any stale interface.
+    indices = list(range(len(ssh_nodes)))
+    mesh = WgMesh(ssh_nodes, c.node_names, remote_bin_dir, c._sudo_prefix(),
+                  iface=WG_IFACE, wg_addresses=_wg_addresses())
+    try:
+        mesh.bring_up(indices)
+    except Exception:
+        mesh.teardown(indices)
+        c.teardown()
+        raise
+    c.wg_mesh = mesh
+    c.wg_mesh_indices = indices
+
+    overrides = {
+        "network": {"wg_enabled": True, "wg_cidr": _WG_CIDR},
+        "cluster": {"enabled": True, "cni": "calico",
+                    "control_plane_node": c.node_names[control_plane_index]},
+    }
+    try:
+        c.start(config_overrides=overrides, agent_as_root=True)
+    except Exception:
+        mesh.teardown(indices)
+        c.teardown()
+        raise
+    return c
+
+
+def _teardown_wg_k0s(c):
+    """Tear down a cluster from _deploy_wg_k0s: daemons first, then k0s, then mesh."""
+    # Ask the controller to bring k0s down (best effort), but do NOT rely on the
+    # async reset completing — force a direct per-node k0s reset below.
+    try:
+        c.k8s_down(reset=True)
+        c.wait_k8s_phase("down", timeout=120)
+    except Exception:
+        pass
+    # Kill spurctld/spurd BEFORE the manual k0s reset: k8s_down(reset=True) kicks
+    # off an async reconcile that stops the k0s components itself, so if the
+    # daemons are still alive while _reset_k0s_all_nodes runs its `k0s reset`/
+    # `rm -rf`, that reconcile races the manual cleanup. Tearing the daemons down
+    # first leaves nothing contending with it.
+    c.teardown()
+    _reset_k0s_all_nodes(c)
+    mesh = getattr(c, "wg_mesh", None)
+    if mesh is not None:
+        mesh.teardown(getattr(c, "wg_mesh_indices", []))
+
+
+@pytest.fixture
+def wg_k0s_cluster(ssh_nodes, remote_bin_dir):
+    """Rootful k0s-over-WireGuard cluster (calico, wg_enabled). Needs >= 3 nodes
+    for an etcd quorum path and rootful sudo. Tears k0s + mesh down on exit."""
+    if len(ssh_nodes) < 3:
+        pytest.skip(f"k0s-over-mesh needs >= 3 nodes for quorum (got {len(ssh_nodes)})")
+    c = _deploy_wg_k0s(ssh_nodes, remote_bin_dir)
+    yield c
+    _teardown_wg_k0s(c)
+
+
+@pytest.fixture
+def wg_login_topology(ssh_nodes, remote_bin_dir):
+    """3-node topology for the login-node reachability scenario:
+
+      node0 = spur controller (spurctld + spurd, mesh head, NOT in k0s scope)
+      node1 = k8s control-plane (spurd, k0s CP)
+      node2 = login node (spurd, meshed, NOT in k0s scope)
+
+    All three are meshed; only node1 is in the k0s scope. Asserts the login node
+    reaches every mesh node.
+    """
+    if len(ssh_nodes) < 3:
+        pytest.skip(f"login-node topology needs 3 nodes (got {len(ssh_nodes)})")
+    c = _deploy_wg_k0s(ssh_nodes, remote_bin_dir, control_plane_index=1)
+    yield c
+    _teardown_wg_k0s(c)
