+++ tests/pytest.ini
@@ -1,6 +1,6 @@
 [pytest]
-pythonpath = native_host/e2e native_host/wg_e2e k8s/e2e
-testpaths = native_host/e2e native_host/wg_e2e k8s/e2e
+pythonpath = native_host/e2e k8s/e2e
+testpaths = native_host/e2e k8s/e2e
 python_files = test_*.py
 python_classes = Test*
 python_functions = test_*
